Move the service specific code into a separate file
